Add files via upload
Kaggle link here: https://www.kaggle.com/rileykin/what-year-did-the-industrial-revolution-start

**Results**

The overall average of all of the commodities deviating from their 2 standard deviation range is 1770 but when we look at the mean and median of the most important commodities we see that they deviated from their 2 standard deviation ranges in 1763 (mean) and 1759 (median). Given the difference of 4 years between the mean and median, I believe the median is more informative to us in this case (Clothing-Stockings appears to be an outlier that doesn't exceed its range until 1790).

Wikipedia's start date for the Industrial Revolution beginning in 1760 is re-confirmed by this analysis! I think it would also be possible to argue the Industrial Revolution began slightly before 1760 as the most relevant of the tracked commodities in this database began to diverge from their 2 standard deviation range around 1759.